chore: split payload into different module
To avoid circular python imports
